requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asn1crypto==1.5.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bitstring==4.0.2
bleach==6.0.0
boto3==1.28.5
botocore==1.31.5
cachetools==5.3.1
catboost==1.2
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
cmake==3.27.7
comm==0.1.3
contextlib2==21.6.0
contourpy==1.1.0
cryptography==40.0.2
cycler==0.11.0
dataclasses-json==0.5.9
deap==1.3.3
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
ecos==2.0.12
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.12.2
fonttools==4.40.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.10.0
geneticalgs==1.0.1
google==3.0.0
google-api-core==2.11.1
google-api-python-client==2.92.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.1
graphviz==0.20.1
gspread==5.10.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
jaraco.classes==3.3.0
jedi==0.18.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.1
jsonpointer==2.4
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
keyring==23.11.0
kiwisolver==1.4.4
langchain==0.0.228
langchainplus-sdk==0.0.20
lightgbm==4.0.0
llvmlite==0.40.1
MarkupSafe==2.1.5
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.1
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.2
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numba==0.57.1
numexpr==2.8.4
numpy==1.24.4
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
oscrypto==1.3.0
osqp==0.6.3
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathos==0.3.1
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.8.1
plotly==5.15.0
pox==0.3.3
ppft==1.7.6.7
prince==0.11.0
prometheus-client==0.17.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.18.0
pydantic==1.10.11
Pygments==2.15.1
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
qdldl==0.1.7.post0
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.29.1
regex==2024.5.15
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.8.10
rpy2==3.5.15
rsa==4.9
s3transfer==0.6.1
sagemaker==2.197.0
schema==0.7.5
scikit-learn==1.3.0
scikit-survival==0.22.1
scipy==1.11.1
seaborn==0.12.2
Send2Trash==1.8.2
shap==0.42.1
six==1.16.0
sklearn-genetic-opt==0.10.1
slicer==0.0.7
smdebug-rulesconfig==1.0.1
sniffio==1.3.0
snowflake-connector-python==3.0.4
snowflake-sqlalchemy==1.4.7
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==1.4.49
stack-data==0.6.2
statsmodels==0.14.0
sympy==1.12
tblib==1.7.0
tenacity==8.2.2
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
toolz==0.12.0
torch==2.1.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
tweedie==0.0.9
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.16
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
widgetsnbextension==4.0.8
wordcloud==1.9.3
xgboost==1.7.6
yarl==1.9.2
zipp==3.16.2
